requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
coverage==7.3.2
crashtest==0.4.1
distlib==0.3.7
dj_email_url==1.0.6
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf_util==1.0.17
drf-yasg==1.21.7
dulwich==0.21.6
environs==9.5.0
factory_boy==3.3.0
faker==19.13.0
filelock==3.13.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.19.2
keyring==24.2.0
Markdown==3.5.1
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.8.0
phonenumbers==8.13.24
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
psycopg==3.1.13
psycopg2-binary==2.9.9
pre-commit==3.5.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.20.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
shellingham==1.5.4
social-auth-app-django==5.4.0
sqlparse==0.4.4
tomlkit==0.12.2
trove-classifiers==2023.11.9
stripe==7.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.6
zipp==3.17.0
